// src/api/__generated__/wallet/model/myRankResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 我的排名
 */
export interface MyRankResponse {
  /** 我的名次，从 1 开始 */
  rank?: number;
  /** 我的累计积分 */
  totalPoints?: number;
  /** 我的名次奖励金额（USDC） */
  rewardAmount?: number;
}